first release 2.0 (now it builds correctly)

+++ lib/data_compressor.cpp
@@ -11,7 +11,7 @@ data_compressor::data_compressor(size_t *freq) : tree(freq) {
 std::vector<code> data_compressor::compress(symbol *data, size_t cnt) {
     auto res = *new std::vector<code>();
     res.reserve(cnt);
-    for (int i = 0; i < cnt; ++i) {
+    for (size_t i = 0; i < cnt; ++i) {
         res.push_back(codes[data[i]]);
     }
     return res;
